Add --add-titles option
When this option is enabled, each parsed note will be considered to
start with a heading:

  # Title-of-note

even when no such line is present in the source file. The title is
inferred based on the filename of the note.

This option makes heavily nested note embeds make more sense in the
exported document, since it shows which note the embedded content comes
from. It loosely matches the behaviour of the mainline Obsidian UI when
viewing notes in preview mode.