Update groupName for pulldown-cmark in renovate config
66108b5128fcb794bfa774ddeca53c6444fefbd0 didn't actually result in a
grouped update for pulldown-cmark and pulldown-cmark-to-cmark, both of
which have available updates and both of which must be upgraded at the
same time in this instance.

Reading the renovate documentation, it is unclear why this configuration
didn't work, but my best guess is that there's some unexpected behavior
when groupName == packageName.